Updated Frontend test based on changes made from previous PR in TestMaster.
