+++ FRONTEND/newFrontEnd/apTime/src/app/components/create-task-dialog/create-task-dialog.component.spec.ts
@@ -4,11 +4,10 @@
  *
  *  Unit Test - Frontend
  */
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, getTestBed, TestBed} from '@angular/core/testing';
 import { MatIconModule} from '@angular/material/icon';
 import { CreateTaskDialogComponent } from './create-task-dialog.component';
 import {CustomMaterialModule} from '../../modules/material.module';
-import {HttpClient, HttpClientModule} from '@angular/common/http';
 import {AmplifyService} from 'aws-amplify-angular';
 import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
 import {RouterTestingModule} from '@angular/router/testing';
@@ -20,24 +19,33 @@ import {
   MatDialogModule,
   MatDialogRef,
   MAT_DIALOG_DATA,
-  MatButtonModule,
-  MatRadioModule,
   MatSelectModule
 } from '@angular/material';
-import {DialogData} from '../tasks/tasks.component';
-import {ConfirmTaskDialogComponent} from '../confirm-task-dialog/confirm-task-dialog.component';
 import {SessionService} from '../../services/session.service';
+import any = jasmine.any;
+import {HttpClientTestingModule, HttpTestingController} from "@angular/common/http/testing";
+
+
 
 describe('CreateTaskDialogComponent', () => {
   let component: CreateTaskDialogComponent;
   let fixture: ComponentFixture<CreateTaskDialogComponent>;
+  let diaMatRef: CreateTaskDialogComponent;
+  let httpMock: HttpTestingController;
+  let mockTest: any;
+  let session: SessionService;
+  let spy = any;
+
+  const dialogMock = {
+    close: () =>{ }
+  };
 
   beforeEach(async(() => {
     TestBed.resetTestEnvironment();
     TestBed.initTestEnvironment(BrowserDynamicTestingModule,
         platformBrowserDynamicTesting());
     TestBed.configureTestingModule({
-      imports: [ HttpClientModule, BrowserAnimationsModule,
+      imports: [ HttpClientTestingModule, BrowserAnimationsModule,
         RouterTestingModule.withRoutes([]),
         MatIconModule,
         MatSelectModule,
@@ -47,9 +55,9 @@ describe('CreateTaskDialogComponent', () => {
         MatDialogModule,
 
       ],
-      providers: [AmplifyService, HttpClient,  {
+      providers: [AmplifyService, {
         provide: MatDialogRef,
-        useValue: {}},
+        useValue: dialogMock,},
         { provide: MAT_DIALOG_DATA, useValue: {}}],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
       declarations: [ CreateTaskDialogComponent, ]
@@ -60,14 +68,41 @@ describe('CreateTaskDialogComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(CreateTaskDialogComponent);
     component = fixture.componentInstance;
+    httpMock = getTestBed().get(HttpTestingController);
     fixture.detectChanges();
   });
 
+  it('should create', () => {
+    expect(component.onNoClick).toBeTruthy();
+    const categoryMineRequest = httpMock.expectOne('http://localhost:8001/category/');
+    categoryMineRequest.flush([]);
+    /*const categoryPublicRequest = httpMock.expectOne('http://localhost:8001/category/public');
+    categoryPublicRequest.flush([]);*/
+    httpMock.verify();
+  });
+
+  it('should have error for failed category requests', () => {
+    expect(component.onNoClick).toBeTruthy();
+    const categoryMineRequest = httpMock.expectOne('http://localhost:8001/category/');
+    categoryMineRequest.error(new ErrorEvent('Generated testing error'));
+    /*const categoryPublicRequest = httpMock.expectOne('http://localhost:8001/category/public');
+    categoryPublicRequest.flush([]);*/
+    httpMock.verify();
+  });
+
+  it('Test onNoClick DialogRef - Close', () => {
+    let spy = spyOn(component.dialogRef, 'close').and.callThrough();
+    component.onNoClick();
+    expect(spy).toHaveBeenCalled();
+    // add expects to test that onNoClick does what it is supposed to
+  });
+  
   it('should create', () => {
     expect(component).toBeTruthy();
   });
 
   it('should onNoclick', () => {
+    component.onNoClick();
     expect(component.onNoClick).toBeTruthy();
   });
 
@@ -76,16 +111,12 @@ describe('CreateTaskDialogComponent', () => {
   });
 
   it('should dateConversion', () => {
+    let date = new Date();
+    let time = '12:00PM';
+    component.dateConversion(time, date);
     expect(component.dateConversion).toBeTruthy();
   });
 
-  /**
-  it('testing pauseTask(number) ', () => {
-    // const testCreate = fixture.debugElement.injector.get(ConfirmTaskDialogComponent);
-    fixture.detectChanges();
-    expect(component.changeMinEndDate).toHaveBeenCalled();
-  });
-*/
   it('should be Defined onNoClick', async(() => {
     spyOn(component, 'onNoClick').and.callThrough();
     fixture.whenStable().then(() => {
@@ -94,23 +125,24 @@ describe('CreateTaskDialogComponent', () => {
     });
   }));
 
-  it('should be Defined onYesClick', async(() => {
-    spyOn(component, 'onYesClick').and.callThrough();
+  it('should be Defined and Should be Called - changeMinEndDate', async(() => {
+    let spy = spyOn(component, 'changeMinEndDate').and.callThrough();
+    component.changeMinEndDate();
     fixture.whenStable().then(() => {
-      expect(component.onYesClick).toBeDefined();
-      expect(component.onYesClick).toHaveBeenCalledTimes(0);
+    expect(spy).toBeDefined();
+    expect(spy).toHaveBeenCalled();
     });
   }));
 
-  it('should be Defined changeMinEndDate', async(() => {
-    spyOn(component, 'changeMinEndDate').and.callThrough();
+  it('should be Defined and Should be Called - dateConversion', async(() => {
+    let date = new Date();
+    let time = '12:00PM';
+    let spy = spyOn(component, 'dateConversion').and.callThrough();
+    component.dateConversion(time, date);
     fixture.whenStable().then(() => {
-      expect(component.changeMinEndDate).toBeDefined();
-      expect(component.changeMinEndDate).toHaveBeenCalled();
+      expect(spy).toBeDefined();
+      expect(spy).toHaveBeenCalled();
     });
   }));
 });
-
-
-
